move convert_to_dataset into mhkit.utils.data_utils
