+++ mhkit/power/characteristics.py
@@ -2,6 +2,7 @@ import pandas as pd
 import xarray as xr
 import numpy as np
 from scipy.signal import hilbert
+from mhkit.utils import convert_to_dataset
 
 def instantaneous_frequency(um, time_dimension="", to_pandas=True):
 
@@ -37,7 +38,7 @@ def instantaneous_frequency(um, time_dimension="", to_pandas=True):
             f'time_dimension must be of type bool. Got: {type(time_dimension)}')
 
     # Convert input to xr.Dataset
-    um = _convert_to_dataset(um, 'data')
+    um = convert_to_dataset(um, 'data')
     
     if time_dimension != '' and time_dimension not in um.coords:
         raise ValueError('time_dimension was supplied but is not a dimension '
@@ -100,8 +101,8 @@ def dc_power(voltage, current, to_pandas=True):
             f'to_pandas must be of type bool. Got: {type(to_pandas)}')
 
     # Convert inputs to xr.Dataset
-    voltage = _convert_to_dataset(voltage, 'voltage')
-    current = _convert_to_dataset(current, 'current')
+    voltage = convert_to_dataset(voltage, 'voltage')
+    current = convert_to_dataset(current, 'current')
 
     # Check that sizes are the same
     if not (voltage.sizes == current.sizes and len(voltage.data_vars) == len(current.data_vars)):
@@ -166,8 +167,8 @@ def ac_power_three_phase(voltage, current, power_factor, line_to_line=False, to_
             f'to_pandas must be of type bool. Got: {type(to_pandas)}')
 
     # Convert inputs to xr.Dataset
-    voltage = _convert_to_dataset(voltage, 'voltage')
-    current = _convert_to_dataset(current, 'current')
+    voltage = convert_to_dataset(voltage, 'voltage')
+    current = convert_to_dataset(current, 'current')
 
     # Check that sizes are the same
     if not len(voltage.data_vars) == 3:
@@ -190,60 +191,3 @@ def ac_power_three_phase(voltage, current, power_factor, line_to_line=False, to_
 
     return P
 
-def _convert_to_dataset(data, name='data'):
-    """
-    Converts the given data to an xarray.Dataset.
-    
-    This function is designed to handle inputs that can be either a pandas DataFrame, a pandas Series,
-    an xarray DataArray, or an xarray Dataset. It ensures that the output is consistently an xarray.Dataset.
-    
-    Parameters
-    ----------
-    data: pandas DataFrame, pandas Series, xarray DataArray, or xarray Dataset
-        The data to be converted. 
-    
-    name: str (Optional)
-        The name to assign to the data variable in case the input is an xarray DataArray without a name.
-        Default value is 'data'.
-    
-    Returns
-    -------
-    xarray.Dataset
-        The input data converted to an xarray.Dataset. If the input is already an xarray.Dataset,
-        it is returned as is.
-    
-    Examples
-    --------
-    >>> df = pd.DataFrame({'A': [1, 2, 3], 'B': [4, 5, 6]})
-    >>> ds = _convert_to_dataset(df)
-    >>> type(ds)
-    <class 'xarray.core.dataset.Dataset'>
-    
-    >>> series = pd.Series([1, 2, 3], name='C')
-    >>> ds = _convert_to_dataset(series)
-    >>> type(ds)
-    <class 'xarray.core.dataset.Dataset'>
-    
-    >>> data_array = xr.DataArray([1, 2, 3])
-    >>> ds = _convert_to_dataset(data_array, name='D')
-    >>> type(ds)
-    <class 'xarray.core.dataset.Dataset'>
-    """
-    if not isinstance(data, (pd.DataFrame, pd.Series, xr.DataArray, xr.Dataset)):
-        raise TypeError("Input data must be of type pandas.DataFrame, pandas.Series, "
-                    "xarray.DataArray, or xarray.Dataset")
-
-    if not isinstance(name, str):
-        raise TypeError("The 'name' parameter must be a string")     
-
-    # Takes data that could be pd.DataFrame, pd.Series, xr.DataArray, or 
-    # xr.Dataset and converts it to xr.Dataset
-    if isinstance(data, (pd.DataFrame, pd.Series)):
-        data = data.to_xarray()
-
-    if isinstance(data, xr.DataArray):
-        if data.name is None:
-            data.name = name # xr.DataArray.to_dataset() breaks if the data variable is unnamed
-        data = data.to_dataset()
-
-    return data
