+++ mhkit/utils/__init__.py
@@ -2,6 +2,6 @@ from .time_utils import matlab_to_datetime, excel_to_datetime
 from .stat_utils import get_statistics, vector_statistics, unwrap_vector, magnitude_phase, unorm
 from .cache import handle_caching, clear_cache
 from .upcrossing import upcrossing, peaks, troughs, heights, periods, custom
-from .data_utils import _convert_to_dataset
+from .data_utils import convert_to_dataset
 
 _matlab = False  # Private variable indicating if mhkit is run through matlab
